requirements.txt
Appium-Python-Client==0.38
PyYAML==3.13
selenium==3.141.0
urllib3==1.24.1
